routing code refactoring
